Move PrintProgramError impls
